All dragonfruit sensors working (#43)
* All dragonfruit sensors working
* Documentation for Dragonfruit Sensors
* All dragonfruit sensors working
* Documentation for Dragonfruit Sensors
* Additional fixes to sensors
* Scaling fix to dragonfruit solar irr sensor
* Dragonfruit sp215 and solar panel fixes
* Pulled Xbee enable pin high
* Working Dragonfruit Solar Irr sensor
* Updated comments and working apogee sp215
* Modified apogee sp215 to properly initialize wire
* Updated comments
* Xbee comments
* Updated xbee documentation

+++ src/gen_dragonfruit/gd_board.h
@@ -19,10 +19,10 @@ struct gd_packet{
   uint32_t uptime_ms;         // Time since start of program
   uint16_t batt_mv;           // Battery Voltage (in milli volts)
   uint16_t panel_mv;          // Panel Voltage (in milli volts)
-  uint16_t apogee_sp215;      // Solar radiance
-  int16_t mpl115a2t1_temp;    // Temperature (Celsius)
+  uint32_t apogee_sp215;      // Solar radiance
+  uint16_t mpl115a2t1_temp;    // Temperature (centiKelvin)
   uint16_t hih6131_humidity_pct;  // Humidity (percentage)
-  uint16_t mpl115a2t1_press;  // Pressure (kPa)
+  uint32_t mpl115a2t1_press;  // Pressure (Pa)
 };
 
 struct gd_heartbeat_packet{
