+++ src/gen_dragonfruit/gd_dev_adafruit_MPL115A2_press.h
@@ -3,5 +3,5 @@
 #ifndef _GD_ADAFRUIT_MPL115A2_PRESS_H
 #define _GD_ADAFRUIT_MPL115A2_PRESS_H
 void gd_dev_adafruit_MPL115A2_press_open(void);
-uint16_t gd_dev_adafruit_MPL115A2_press_read(void);
+uint32_t gd_dev_adafruit_MPL115A2_press_read(void);
 #endif
