+++ src/gen_dragonfruit/gd_dev_adafruit_MPL115A2_temp.h
@@ -3,5 +3,5 @@
 #ifndef _GD_ADAFRUIT_MPL115A2_TEMP_H
 #define _GD_ADAFRUIT_MPL115A2_TEMP_H
 void gd_dev_adafruit_MPL115A2_temp_open(void);
-int32_t gd_dev_adafruit_MPL115A2_temp_read(void);
+uint16_t gd_dev_adafruit_MPL115A2_temp_read(void);
 #endif
